Creates a new notes.py file to store lecture notes
